fix(samples,hal): timing-safe VRAM uploads and full VRAM init for gb-3d
On accurate emulators (mGBA) the cube showed banding: the per-byte
poll-STAT-then-write upload can land a write inside mode 3, where
hardware silently drops it. The double-buffered sample now uploads the
whole 1920-byte page via GDMA inside vblank on CGB (new
Cgb.CopyToVram HAL wrapper; ~3840 dots, fits the 4560-dot window) and
flips pages by toggling LCDC bit 4 (new Lcd.SelectTileData) instead of
rewriting 120 tilemap cells; DMG falls back to a bounded
bytes-per-vblank CPU copy. All three samples now clear the full
tilemap and unowned tile data at init, so boot-logo remnants and
tile-0 replication cannot show through. Also fixes TileData.Clear/
SetRow computing tile*16 in byte width - it wrapped mod 256 and
silently cleared the wrong tile for any index >= 16 (exposed by the
DMG hand-off logo; Tilemap gets a Clear helper with the same 16-bit
widening). A new Mode3WriteGuard in the verify harness asserts zero
CPU VRAM writes ever land during mode 3, and the harness now also
asserts the cube animates at the expected rate per mode.

+++ src/Koh.GameBoy/Hal/TileData.cs
@@ -4,10 +4,11 @@ namespace Koh.GameBoy;
 /// typed view over that VRAM region so callers build tiles a row at a time instead of poking bytes.</summary>
 public static unsafe class TileData
 {
-    /// <summary>Zero every pixel of a tile (all colour 0).</summary>
+    /// <summary>Zero every pixel of a tile (all colour 0). The offset is widened to 16-bit before the
+    /// multiply (tile 16 -> 256 would overflow a byte and clear the wrong tile's data).</summary>
     public static void Clear(byte tile)
     {
-        byte* p = Gb.Vram + tile * 16;
+        byte* p = Gb.Vram + (ushort)((ushort)tile * 16);
         for (byte i = 0; i < 16; i++)
             *(p + i) = 0;
     }
@@ -15,7 +16,7 @@ public static unsafe class TileData
     /// <summary>Write one 8-pixel row of a tile as its two bit-planes (low, high).</summary>
     public static void SetRow(byte tile, byte row, byte low, byte high)
     {
-        byte* p = Gb.Vram + tile * 16 + row * 2;
+        byte* p = Gb.Vram + (ushort)((ushort)tile * 16 + row * 2);
         *p = low;
         *(p + 1) = high;
     }
